Divide functionality in to sub components
