+++ src/App.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import './App.css';
 import { Route, Switch } from 'react-router-dom';
-import Chat from './components/chat';
+import ChatContainer from './components/chatContainer';
 
 function App() {
   return (
     <div className="App">
       <Switch>
-        <Route path="/" component={Chat}/>
-        <Route path="/chat" component={Chat}/>
+        <Route path="/" component={ChatContainer}/>
+        <Route path="/chat" component={ChatContainer}/>
       </Switch>
     </div>
   );
