+++ src/components/chatHeader.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ChatHeader = () => {
+    return ( 
+        <h3 style={{fontWeight: '500'}}>
+            Carla Bot
+        </h3>
+    );
+}
+ 
+export default ChatHeader;
